create months by profile completed

+++ src/expenses/expenses.controller.ts
@@ -3,13 +3,15 @@ import { ExpensesService } from './expenses.service';
 import { Expense, Prisma } from '@prisma/client'
 import { PrismaRelations } from 'src/prisma/relations.service';
 import { MonthsService } from '../months/months.service';
+import { ExpensesCalculatorService } from 'src/months/expenseMonthCalculator.service';
 
 @Controller('profile/:profileId/expenses')
 export class ExpensesController {
   constructor(
     private readonly expensesService: ExpensesService, 
     private relation: PrismaRelations,
-    private monthService: MonthsService
+    private monthService: MonthsService,
+    private expensesCalcService: ExpensesCalculatorService
   ) {}
 
   @Post('/')
@@ -21,6 +23,10 @@ export class ExpensesController {
     const expenseCreation = await this.expensesService.createExpense(createExpense); 
     
     await this.relation.addExpenseToActualMonth(expenseCreation, profileIdNumber, monthId)
+
+    const expenses = await this.expensesService.findAll(profileIdNumber)
+
+    await this.expensesCalcService.updateMonthRepository(expenses, monthId)
 
     return expenseCreation
   }
@@ -35,23 +41,54 @@ export class ExpensesController {
   @Get('/:id')
   async findOne(@Param('profileId') profileId: string, @Param('id') id: string) {
     const idNumber = Number(id)
+    const profileIdNumber = Number(profileId)
 
-    const expense = await this.expensesService.findOne(idNumber) 
+    const expense = await this.expensesService.findOne(idNumber, profileIdNumber) 
 
     return expense || 'Este Expense Não Existe'
   }
 
   @Put('/:id')
-  async update(@Param('profileId') profileId: string, @Param('id') id: string, @Body() expenseUpdated: Prisma.ExpenseUpdateInput) {
+  async update(@Param('profileId') profileId: string, @Param('id') id: string, @Body() expenseUpdate: Prisma.ExpenseUpdateInput) {
     const idNumber = Number(id)
+    const profileIdNumber = Number(profileId)
 
-    return await this.expensesService.update(idNumber, expenseUpdated);
+    const expense = await this.expensesService.findOne(idNumber, profileIdNumber) 
+
+    if (!expense) return 'Esse Expense Não Existe'
+    
+    const expenseUpdated = await this.expensesService.update(idNumber, expenseUpdate) 
+
+    const {id: monthId} = await this.monthService.createMonth(profileIdNumber)
+
+    const expenses = await this.expensesService.findAll(profileIdNumber)
+
+    await this.expensesCalcService.updateMonthRepository(expenses, monthId)
+
+    return expenseUpdated;
   }
 
   @Delete('/:id')
-  remove(@Param('id') id: string) {
+  async remove(@Param('profileId') profileId: string, @Param('id') id: string) {
     const idNumber = Number(id)
+    const profileIdNumber = Number(profileId)
 
-    return this.expensesService.remove(idNumber);
+    const expense = await this.expensesService.findOne(idNumber, profileIdNumber) 
+
+    if (!expense) return 'Esse Expense Não Existe'
+
+    const {id: monthId} = await this.monthService.createMonth(profileIdNumber)
+    
+    await this.expensesService.remove(idNumber) 
+
+    const expenses = await this.expensesService.findAll(profileIdNumber)
+
+    if (!expenses.length) {
+      await this.monthService.deleteMonth(monthId)
+    } else {
+      await this.expensesCalcService.updateMonthRepository(expenses, monthId)
+    }
+
+    return `${id} foi excluido`;
   }
 }
